Adds run command
