+++ src/commands/mod.rs
@@ -1,3 +1,4 @@
 pub mod list;
+pub mod run;
 pub mod save;
 pub mod show;
